further split up rendering.go, added more test cases, removed old test artifacts
