+++ pkg/nftables/rendering.go
@@ -4,7 +4,6 @@ import (
 	"bytes"
 	"fmt"
 	"io/ioutil"
-	"net"
 	"net/http"
 	"strings"
 	"text/template"
@@ -24,18 +23,15 @@ type firewallRenderingData struct {
 }
 
 func newFirewallRenderingData(f *Firewall) (*firewallRenderingData, error) {
-	ingress, egress := []string{}, []string{}
+	ingress, egress := nftablesRules{}, nftablesRules{}
 	for _, np := range f.clusterwideNetworkPolicies.Items {
-		if len(np.Spec.Egress) > 0 {
-			egress = append(egress, egressForNetworkPolicy(np)...)
-		}
-		if len(np.Spec.Ingress) > 0 {
-			ingress = append(ingress, ingressForNetworkPolicy(np)...)
-		}
+		i, e := clusterwideNetworkPolicyRules(np)
+		ingress = append(ingress, i...)
+		egress = append(egress, e...)
 	}
 
 	for _, svc := range f.services.Items {
-		ingress = append(ingress, ingressForService(svc)...)
+		ingress = append(ingress, serviceRules(svc)...)
 	}
 
 	snatRules, err := snatRules(f)
@@ -53,8 +49,8 @@ func newFirewallRenderingData(f *Firewall) (*firewallRenderingData, error) {
 		PrivateVrfID:     f.primaryPrivateNet.Vrf,
 		InternalPrefixes: strings.Join(f.spec.InternalPrefixes, ", "),
 		ForwardingRules: forwardingRules{
-			Egress:  uniqueSorted(egress),
-			Ingress: uniqueSorted(ingress),
+			Ingress: ingress,
+			Egress:  egress,
 		},
 		RateLimitRules: rateLimitRules(f),
 		SnatRules:      snatRules,
@@ -102,90 +98,4 @@ func (d *firewallRenderingData) readTpl() (string, error) {
 		return "", err
 	}
 	return string(bytes), nil
-}
-
-type snatRule struct {
-	sourceNetworks string
-	oifname        string
-	to             string
-	comment        string
-}
-
-func (s *snatRule) String() string {
-	return fmt.Sprintf(`ip saddr { %s } oifname "%s" counter snat %s comment "%s"`, s.sourceNetworks, s.oifname, s.to, s.comment)
-}
-
-// snatRules generates the nftables rules for SNAT based on the firewall spec
-func snatRules(f *Firewall) (nftablesRules, error) {
-	if f.primaryPrivateNet == nil {
-		return nil, fmt.Errorf("no primary private network found")
-	}
-
-	rules := nftablesRules{}
-	for _, s := range f.spec.Snat {
-		n, there := f.networkMap[s.Network]
-		if !there {
-			return nil, fmt.Errorf("network not found")
-		}
-
-		hmap := []string{}
-		for k, i := range s.IPs {
-			ip := net.ParseIP(i)
-			if ip == nil {
-				return nil, fmt.Errorf("could not parse ip %s", i)
-			}
-
-			innets := false
-			for _, prefix := range n.Prefixes {
-				_, cidr, err := net.ParseCIDR(prefix)
-				if err != nil {
-					return nil, fmt.Errorf("could not parse cidr %s", prefix)
-				}
-
-				if cidr.Contains(ip) {
-					innets = true
-					break
-				}
-			}
-
-			if !innets {
-				return nil, fmt.Errorf("ip %s is not in any prefix of network %s", i, s.Network)
-			}
-			hmap = append(hmap, fmt.Sprintf("%d : %s", k, ip.String()))
-		}
-
-		var to string
-		if len(s.IPs) == 0 {
-			return nil, fmt.Errorf("need to specify at least one address for SNAT")
-		} else if len(s.IPs) == 1 {
-			to = s.IPs[0]
-		} else {
-			to = fmt.Sprintf("to jhash ip daddr . tcp sport mod %d map { %s }", len(s.IPs), strings.Join(hmap, ", "))
-		}
-
-		snatRule := snatRule{
-			comment:        fmt.Sprintf("snat for %s", s.Network),
-			sourceNetworks: strings.Join(f.primaryPrivateNet.Prefixes, ", "),
-			oifname:        fmt.Sprintf("vlan%d", n.Vrf),
-			to:             to,
-		}
-		rules = append(rules, snatRule.String())
-	}
-	return rules, nil
-}
-
-// rateLimitRules generates the nftables rules for rate limiting networks based on the firewall spec
-func rateLimitRules(f *Firewall) nftablesRules {
-	rules := nftablesRules{}
-	for _, l := range f.spec.RateLimits {
-		n, ok := f.networkMap[l.Network]
-		if !ok {
-			continue
-		}
-		if n.Underlay {
-			continue
-		}
-		rules = append(rules, fmt.Sprintf(`meta iifname "%s" limit rate over %d mbytes/second counter name drop_ratelimit drop`, fmt.Sprintf("vrf%d", n.Vrf), l.Rate))
-	}
-	return rules
 }
