--- pkg/nftables/test_data/case1/expected.nftablev4
@@ -1,25 +1,0 @@
-table ip firewall {
-	chain forward {
-		type filter hook forward priority 1; policy drop;
-
-		# state dependent rules
-		ct state established,related counter accept comment "accept established connections"
-		ct state invalid counter drop comment "drop packets with invalid ct state"
-
-		# icmp
-		ip protocol icmp icmp type echo-request limit rate over 10/second burst 4 packets counter drop comment "drop ping floods"
-		ip protocol icmp icmp type { destination-unreachable, router-solicitation, router-advertisement, time-exceeded, parameter-problem } counter accept comment "accept icmp"
-
-		# dynamic ingress rules
-		ip saddr { 192.168.0.0/24, 192.168.2.0/24 } ip daddr { 212.37.83.1 } tcp dport { 80, 53 } counter accept comment "accept traffic for k8s service test-ns/s1"
-		ip saddr { 212.1.1.1/32 } ip daddr { 212.37.83.2 } tcp dport { 443 } counter accept comment "accept traffic for k8s service test-ns/s2"
-
-		# dynamic egress rules
-		ip daddr { 1.1.1.1/32, 1.0.0.1/32 } tcp dport { 53 } counter accept comment "accept traffic for np np-egress-dns tcp"
-		ip daddr { 1.1.1.1/32, 1.0.0.1/32 } udp dport { 53 } counter accept comment "accept traffic for np np-egress-dns udp"
-		ip daddr { 162.159.200.1/32 } udp dport { 123 } counter accept comment "accept traffic for np np-egress-ntp udp"
-
-		counter comment "count dropped packets"
-		limit rate 10/second counter packets 1 bytes 40 log prefix "nftables-firewall-dropped: "
-	}
-}
