--- pkg/nftables/test_data/case1/policies/np-egress-dns.yaml
@@ -1,21 +1,0 @@
----
-apiVersion: networking.k8s.io/v1
-kind: NetworkPolicy
-metadata:
-  name: np-egress-dns
-  namespace: default
-spec:
-  podSelector: {}
-  policyTypes:
-  - Egress
-  egress:
-  - to:
-    - ipBlock:
-        cidr: 1.1.1.1/32
-    - ipBlock:
-        cidr: 1.0.0.1/32
-    ports:
-    - protocol: UDP
-      port: 53
-    - protocol: TCP
-      port: 53
