--- pkg/nftables/test_data/case1/policies/np-egress-ntp.yaml
@@ -1,17 +1,0 @@
----
-apiVersion: networking.k8s.io/v1
-kind: NetworkPolicy
-metadata:
-  name: np-egress-ntp
-  namespace: default
-spec:
-  podSelector: {}
-  policyTypes:
-  - Egress
-  egress:
-  - to:
-    - ipBlock:
-        cidr: 162.159.200.1/32
-    ports:
-    - protocol: UDP
-      port: 123
