--- pkg/nftables/test_data/case1/services/s1.yaml
@@ -1,20 +1,0 @@
-  apiVersion: v1
-  kind: Service
-  metadata:
-    name: s1
-    namespace: test-ns
-  spec:
-    type: LoadBalancer
-    loadBalancerIP: 212.37.83.1
-    loadBalancerSourceRanges:
-    - 192.168.0.0/24
-    - 192.168.2.0/24
-    ports:
-    - name: http
-      protocol: TCP
-      port: 80
-      targetPort: 8063
-    - name: test
-      protocol: TCP
-      port: 53
-      targetPort: 8064
