--- pkg/nftables/test_data/case1/services/s2.yaml
@@ -1,14 +1,0 @@
-  apiVersion: v1
-  kind: Service
-  metadata:
-    name: s2
-    namespace: test-ns
-  spec:
-    type: LoadBalancer
-    loadBalancerIP: 212.37.83.2
-    loadBalancerSourceRanges:
-    - 212.1.1.1/32
-    ports:
-    - protocol: TCP
-      port: 443
-      targetPort: 10420
